chore: move AutoEmissionTemplateCache file
